fix: 合并同类头文件
